refactor sentence splitting
